refactor(desktop): extract optimistic chat message helpers
