+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatMastraPane/ChatMastraInterface/ChatMastraInterface.tsx
@@ -1,7 +1,6 @@
 import { chatServiceTrpc } from "@superset/chat/client";
 import {
 	chatMastraServiceTrpc,
-	type UseMastraChatDisplayReturn,
 	useMastraChatDisplay,
 } from "@superset/chat-mastra/client";
 import {
@@ -27,62 +26,16 @@ import { McpControls } from "./components/McpControls";
 import { useMcpUi } from "./hooks/useMcpUi";
 import type { ChatMastraInterfaceProps } from "./types";
 import {
+	hasMatchingUserMessage,
+	type MastraHistoryMessage,
+	toOptimisticUserMessage,
+} from "./utils/optimisticUserMessage";
+import {
 	type ChatSendMessageInput,
 	sendMessageForSession,
 	toSendFailureMessage,
 } from "./utils/sendMessage";
 import { toMastraImages } from "./utils/toMastraImages";
-
-type MastraHistoryMessage = NonNullable<
-	UseMastraChatDisplayReturn["messages"]
->[number];
-
-function toOptimisticUserMessage(
-	input: ChatSendMessageInput,
-): MastraHistoryMessage | null {
-	const text = input.payload.content.trim();
-	const images = input.payload.images ?? [];
-	if (!text && images.length === 0) return null;
-
-	return {
-		id: `optimistic-${crypto.randomUUID()}`,
-		role: "user",
-		content: [
-			...(text ? [{ type: "text", text }] : []),
-			...images.map((image) => ({
-				type: "image",
-				data: image.data,
-				mimeType: image.mimeType,
-			})),
-		],
-		createdAt: new Date(),
-	} as MastraHistoryMessage;
-}
-
-function toUserMessageSignature(message: MastraHistoryMessage): string | null {
-	if (message.role !== "user") return null;
-	return message.content
-		.map((part) => {
-			if (part.type === "text") return `text:${part.text}`;
-			if (part.type === "image") return `image:${part.mimeType}:${part.data}`;
-			return `${part.type}:${JSON.stringify(part)}`;
-		})
-		.join("||");
-}
-
-function hasMatchingUserMessage({
-	messages,
-	candidate,
-}: {
-	messages: MastraHistoryMessage[];
-	candidate: MastraHistoryMessage;
-}): boolean {
-	const signature = toUserMessageSignature(candidate);
-	if (!signature) return false;
-	return messages.some(
-		(message) => toUserMessageSignature(message) === signature,
-	);
-}
 
 function useAvailableModels(): {
 	models: ModelOption[];
@@ -343,7 +296,7 @@ export function ChatMastraInterface({
 		if (!pendingImmediateUserMessage) return;
 		if (
 			hasMatchingUserMessage({
-				messages: messages as MastraHistoryMessage[],
+				messages,
 				candidate: pendingImmediateUserMessage,
 			})
 		) {
@@ -355,7 +308,7 @@ export function ChatMastraInterface({
 		if (!pendingImmediateUserMessage) return messages;
 		if (
 			hasMatchingUserMessage({
-				messages: messages as MastraHistoryMessage[],
+				messages,
 				candidate: pendingImmediateUserMessage,
 			})
 		) {
